requirements.txt
flask
flask-cors
gunicorn
flask-mail==0.9.1
python-dotenv==1.0.0 